Rename examples button component to kebab-case
