+++ packages/examples/src/components/examples/FilesExample/CategoriesSection.tsx
@@ -1,7 +1,7 @@
 import type { LucideIcon } from 'lucide-react';
 import { FileIcon, ImageIcon, MusicIcon, VideoIcon } from 'lucide-react';
 
-import { Button } from '@/components/ui/Button';
+import { Button } from '@/components/ui/button';
 
 const CATEGORIES = [
   { id: 'documents', label: 'Documents', icon: FileIcon },
