+++ packages/examples/src/components/examples/FilesExample/FileList.tsx
@@ -7,7 +7,7 @@ import {
   VideoIcon,
 } from 'lucide-react';
 
-import { Button } from '@/components/ui/Button';
+import { Button } from '@/components/ui/button';
 import {
   Table,
   TableBody,
