+++ packages/examples/src/components/examples/FilesExample/UsedStorageCard.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ui/Button';
+import { Button } from '@/components/ui/button';
 import {
   Card,
   CardContent,
