+++ packages/examples/src/components/examples/SettingsExample/NotificationsTab.tsx
@@ -2,7 +2,7 @@ import { type ComponentPropsWithoutRef, forwardRef, useState } from 'react';
 import type { LucideIcon } from 'lucide-react';
 import { AtSignIcon, BellIcon, EyeOffIcon } from 'lucide-react';
 
-import { Button } from '@/components/ui/Button';
+import { Button } from '@/components/ui/button';
 import {
   CardContent,
   CardDescription,
