+++ packages/examples/src/components/examples/StepsExample/FinishStep.tsx
@@ -5,7 +5,7 @@ import {
 } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { Button } from '@/components/ui/Button';
+import { Button } from '@/components/ui/button';
 import {
   CardDescription,
   CardFooter,
