+++ packages/examples/src/components/examples/StepsExample/InterestsStep.tsx
@@ -8,7 +8,7 @@ import {
   MusicIcon,
 } from 'lucide-react';
 
-import { Button } from '@/components/ui/Button';
+import { Button } from '@/components/ui/button';
 import {
   CardContent,
   CardDescription,
