+++ packages/examples/src/components/ui/button/index.ts
@@ -1,2 +1,2 @@
-export * from './Button';
+export * from './button';
 export * from './variants';
