+++ packages/examples/src/routes/Root.tsx
@@ -8,7 +8,7 @@ import {
   SettingsIcon,
 } from 'lucide-react';
 
-import { Button } from '@/components/ui/Button';
+import { Button } from '@/components/ui/button';
 
 const links = [
   {
