+++ packages/examples/src/routes/SharedAxisExample.tsx
@@ -8,7 +8,7 @@ import { TransitionSwitchItem } from 'react-transition-switch';
 import { ChevronLeftIcon, ChevronRightIcon } from 'lucide-react';
 import { toWords } from 'number-to-words';
 
-import { Button } from '@/components/ui/Button';
+import { Button } from '@/components/ui/button';
 import { Checkbox } from '@/components/ui/Checkbox';
 import { FormItem } from '@/components/ui/Form';
 import { Label } from '@/components/ui/Label';
